+++ pkg/cloudcertificates/datamodel.go
@@ -8,8 +8,8 @@ import (
 	"slices"
 	"time"
 
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v12/internal/texts"
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v12/pkg/edgegriderr"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v13/internal/texts"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v13/pkg/edgegriderr"
 	validation "github.com/go-ozzo/ozzo-validation/v4"
 )