+++ pkg/session/session.go
@@ -9,8 +9,8 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v12/pkg/edgegrid"
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v12/pkg/log"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v13/pkg/edgegrid"
+	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v13/pkg/log"
 )
 
 type (
@@ -64,7 +64,7 @@ var (
 
 const (
 	// Version is the client version
-	Version = "12.0.0"
+	Version = "13.0.0"
 )
 
 // New returns a new session